+++ tests/test_no_parellel.py
@@ -1,6 +1,12 @@
+import tempfile
+import textwrap
 import time
 
+from click import testing as cli_testing
+
 import sky
+from sky import cli
+from sky import exceptions
 from sky.utils import registry
 
 
@@ -27,3 +33,64 @@ def test_optimize_speed(enable_all_clouds):
         sky.Resources(cpus='4+', memory='4+', accelerators='A100-80GB:8'))
     _test_optimize_speed(
         sky.Resources(cpus='4+', memory='4+', accelerators='tpu-v3-32'))
+
+
+class TestAllCloudsEnabled:
+
+    def test_accelerator_mismatch(self, enable_all_clouds):
+        """Test the specified accelerator does not match the instance_type."""
+
+        spec = textwrap.dedent("""\
+            resources:
+                cloud: aws
+                instance_type: p3.2xlarge""")
+        cli_runner = cli_testing.CliRunner()
+
+        def _capture_mismatch_gpus_spec(file_path, gpus: str):
+            result = cli_runner.invoke(cli.launch,
+                                       [file_path, '--gpus', gpus, '--dryrun'])
+            assert isinstance(result.exception,
+                              exceptions.ResourcesMismatchError)
+            assert 'Infeasible resource demands found:' in str(result.exception)
+
+        def _capture_match_gpus_spec(file_path, gpus: str):
+            result = cli_runner.invoke(cli.launch,
+                                       [file_path, '--gpus', gpus, '--dryrun'])
+            assert not result.exit_code
+
+        with tempfile.NamedTemporaryFile('w', suffix='.yml') as f:
+            f.write(spec)
+            f.flush()
+
+            _capture_mismatch_gpus_spec(f.name, 'T4:1')
+            _capture_mismatch_gpus_spec(f.name, 'T4:0.5')
+            _capture_mismatch_gpus_spec(f.name, 'V100:2')
+            _capture_mismatch_gpus_spec(f.name, 'v100:2')
+            _capture_mismatch_gpus_spec(f.name, 'V100:0.5')
+
+            _capture_match_gpus_spec(f.name, 'V100:1')
+            _capture_match_gpus_spec(f.name, 'V100:1')
+            _capture_match_gpus_spec(f.name, 'V100')
+
+    def test_k8s_alias(self, enable_all_clouds):
+        cli_runner = cli_testing.CliRunner()
+
+        result = cli_runner.invoke(cli.launch, ['--cloud', 'k8s', '--dryrun'])
+        assert not result.exit_code
+
+        result = cli_runner.invoke(cli.launch,
+                                   ['--cloud', 'kubernetes', '--dryrun'])
+        assert not result.exit_code
+
+        result = cli_runner.invoke(cli.launch,
+                                   ['--cloud', 'notarealcloud', '--dryrun'])
+        assert isinstance(result.exception, ValueError)
+
+        result = cli_runner.invoke(cli.show_gpus, ['--cloud', 'k8s'])
+        assert not result.exit_code
+
+        result = cli_runner.invoke(cli.show_gpus, ['--cloud', 'kubernetes'])
+        assert not result.exit_code
+
+        result = cli_runner.invoke(cli.show_gpus, ['--cloud', 'notarealcloud'])
+        assert isinstance(result.exception, ValueError)
